python/requirements.txt
# Python dependencies for Restaurant Management API
# Core FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database and ORM
sqlmodel>=0.0.14
sqlalchemy>=2.0.0
pydantic>=2.4.0
pydantic_settings

# Authentication and Security
python-jose[cryptography]>=3.3.0  # For JWT tokens
passlib[bcrypt]>=1.7.4            # For password hashing
bcrypt>=4.0.0                     # Password hashing backend

# Form handling
python-multipart>=0.0.6

# Email validation
email-validator>=2.0.0

# Database drivers (choose based on your database)
# SQLite (included in Python standard library by default)
# psycopg2-binary>=2.9.0          # For PostgreSQL
# PyMySQL>=1.0.0                  # For MySQL/MariaDB

# Development and testing dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0                     # For testing FastAPI endpoints

# Linting and formatting (optional)
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Environment variables
python-dotenv>=1.0.0

# For QR code generation (optional)
qrcode[pil]>=7.4.2

# For file uploads and image processing (optional)
pillow>=10.0.0

# For Redis caching (optional)
redis>=5.0.0

# For background tasks (optional)  
celery>=5.3.0
psycopg2